Convert 'measure' meaning to 'metric' and vice versa
