+++ mllib/src/main/scala/org/apache/spark/mllib/linalg/distance/SquaredEuclideanDistanceMeasure.scala
@@ -17,26 +17,27 @@
 
 package org.apache.spark.mllib.linalg.distance
 
-import org.apache.spark.mllib.linalg.Vectors
+import org.apache.spark.annotation.Experimental
+import org.apache.spark.mllib.linalg
 
-class CosineDistanceMetricSuite extends GeneralDistanceMetricSuite {
-  override def distanceFactory = new CosineDistanceMetric
+/**
+ * :: Experimental ::
+ * Squared euclidean distance implementation
+ */
+@Experimental
+class SquaredEuclideanDistanceMeasure extends DistanceMeasure{
 
-  test("concreate distance check") {
-    val vector1 = Vectors.dense(1.0, 2.0)
-    val vector2 = Vectors.dense(3.0, 4.0)
-    val distance = distanceFactory(vector1, vector2)
+  /**
+   * Calculates the squared euclidean distance between 2 points
+   *
+   * @param v1 a Vector defining a multidimensional point in some feature space
+   * @param v2 a Vector defining a multidimensional point in some feature space
+   * @return a scalar doubles of the distance
+   */
+  override def apply(v1: linalg.Vector, v2: linalg.Vector): Double = {
+    validate(v1, v2)
 
-    val coefficient = 100000
-    Math.floor(distance * coefficient) / coefficient should be(0.01613)
-  }
-
-  test("two vectors have the same magnitude") {
-    val vector1 = Vectors.dense(1.0, 1.0)
-    val vector2 = Vectors.dense(2.0, 2.0)
-    val distance = distanceFactory(vector1, vector2)
-
-    val coefficient = 100000
-    Math.floor(distance * coefficient) / coefficient should be(0.0)
+    val diffVector = (v1.toBreeze - v2.toBreeze)
+    diffVector.dot(diffVector)
   }
 }
