+++ mllib/src/test/scala/org/apache/spark/mllib/linalg/distance/ManhattanDistanceMetricSuite.scala
@@ -17,21 +17,6 @@
 
 package org.apache.spark.mllib.linalg.distance
 
-import org.apache.spark.mllib.linalg.Vectors
-
-class MinkowskiDistanceMeasureSuite extends GeneralDistanceMeasureSuite {
-  override def distanceFactory: DistanceMeasure = new MinkowskiDistanceMeasure(4.0)
-
-  test("the distance between the vectors should be expected") {
-    val vector1 = Vectors.dense(0, 0, 0)
-    val vector2 = Vectors.dense(2, 3, 4)
-
-    val measure = new MinkowskiDistanceMeasure
-    assert(measure.exponent == 3.0, s"the default value for exponent should be ${measure.exponent}")
-
-    val distance = measure(vector1, vector2)
-    val roundDistance = GeneralDistanceMeasureSuite.roundValue(distance, 10)
-    val expected = 4.6260650092
-    assert(roundDistance == expected, s"the distance between the vectors should be ${expected}")
-  }
+class ManhattanDistanceMetricSuite extends GeneralDistanceMetricSuite {
+  override def distanceFactory = new ManhattanDistanceMetric()
 }
