+++ mllib/src/test/scala/org/apache/spark/mllib/linalg/distance/WeightedEuclideanDistanceMetricSuite.scala
@@ -17,16 +17,12 @@
 
 package org.apache.spark.mllib.linalg.distance
 
-import org.apache.spark.annotation.Experimental
+import org.apache.spark.mllib.linalg.Vectors
 
-/**
- * :: Experimental ::
- * This trait is used for objects which can determine a distance metric between two points
- *
- * 1. d(x, y) >= 0 (non-negative)
- * 2. d(x, y) = 0 if and only if x = y (identity of indiscernibles)
- * 3. d(x, y) = d(y, x) (symmetry)
- * 4. d(x, z) <= d(x, y) + d(y, z) (triangle inequality)
- */
-@Experimental
-trait DistanceMetric extends DistanceMeasure
+class WeightedEuclideanDistanceMetricSuite extends GeneralDistanceMetricSuite {
+
+  override def distanceFactory: DistanceMetric = {
+    val weights = Vectors.dense(0.1, 0.1, 0.1, 0.1, 0.1, 0.1) // size should be 6
+    new WeightedEuclideanDistanceMetric(weights)
+  }
+}
